Relax condition on unknown HITRAN quantum numbers
Instead of aborting with a runtime error when encountering an unknown
quantum number, continue and set it to 0/0. The format of QNs in Hitran is
extended/changed regularly and very difficult to keep up with.

Temporary fix to be able to read the latest HITRAN online version. Proper fix
according to Richard is to use the named upper and lower states quantum
numbers which are provided as an additional field. The current mechanism
should be replaced by that.