Add lbl_checked variable to act as yet another checked variable
Also removes lm_p_lim since it is not used anywhere anymore

The new variable is accompanied by lbl_checkedCalc which presently
just moves a few of the checks we had before per level to outside.
The main point of this new test is to ensure that the absorption
lines are defined in a way that satisfies the requirement of the
functions that need abs_lines_per_species as input.  For now, I only
really check Zeeman being normal.  I will add NLTE checks in the
future as needs become necessary.

WARNING:  This breaks a lot of controlfiles.  The solution is that
wherever you defined lm_p_lim yourself remove that line, and add a
call to lbl_checkedCalc() before your call to any or the first
method that requires line-by-line calculations