Fix reference comparisons in TestNLTE
The test was storing the current result with the same filenames as the
reference files, reading them for comparison instead of the real reference
files. Thus, the test never failed.

Added 'REFERENCE' to the filenames and updated them with current values.
The test was failing with the old reference files, I assume they were
outdated.