Change to use LineShape::SingleSpeciesModel for MPM2020 code
Also, added basic Zeeman stuff and line shape parameter code
but leave it commented out for now.  The former requires changing
the interface to the function slightly, and I do not know how to
make it work without also making the interface happen through a
propmat_clearsky-type interface.  The latter will be addressed in
a coming commit.