Fix bug in abs_lines_per_speciesCreateFromLines
The contents of the WSV was not emptied. When the WSM was called
multiple times in the same controlfiles, this led to duplicate lines
being present in abs_lines_per_species.

The bug was only showing in Release mode, because output-only variables are
zeroed before use in Debug mode.

This fixes all failing lineshape tests in Release mode.