Now the Full diagonalization works (with overflow/underflow!)
The results are mostly looking very good right now but for some
reason there is an error that is probably caused by some division
near zero.  I do not know why the Fortran code does not experience
the same issues.

The code is as it stands 3 times faster than the Fortran code when
run with Release flag.  When run with Debug flag the code is instead
10 times slower than the Fortran code