Remove offending overflow/underflow villain
There is some issue with the bounds between high/low pressure close
to the line centers in the current method.  The code is still in the
text but commented out.  It seems to be removed completely in newer
versions of this type of software from the HITRAN people.  In any
case, the code now runs without any issues at high pressures.  I
will get to writing an interface that allows selecting limiting
pressure.

In short: must use Lorentz because the code does not want to work
well at low pressures