Some minor restructure
Next goal is to make it work with normal ARTS input (SI)

Also, it might be that the speedup is because openblas is cheating
with the number of cores it uses.  You can't seem to control it,
but it makes the code run faster.  (I compiled openblas myself
and it makes a 3min10sec -> 2min40sec diff in my code, but I did
not find how to turn off the multithreading)