Add Group to the list of ARTS namespaces for the CPP API
This addition is basically not really necessary but it makes it so
that when Modules comes to C++20, the autoarts.h file can be used
alone to define a proper public interface.  (As it is today, if
you just include autoarts.h and link to the public_arts_interface
target, you actually include quite a large subset of ARTS's names,
so the interface is "public".  It would be nice to not populate
the global namespace this much but have a "private" interface so
that only things in declared specifically in autoarts.h and nothing
else populates the names of the program including it.)