Full cyclic Lagrange interpolation
With this a LagrangeType::Cyclic now works at any cyclicity
(0, 360), (-180, 180), (0, 2*pi), (0, 0.5), (-0.123, 0.456)

Note that the speed are similar as in the previous case but
that this comes with one cost for cyclic interpolations.  If
the cyclic interpolation points include the crossings, such
that the original x grid contains 0, ..., 359, 360,
and the point of interest is at 359.5, then only linear
interpolation between 359 and 360 is used.  Of course,
it is up to the user of the algorithm to make sure that
y is the same in both 0 and 360, otherwise using the
function gives undefined output.