Remove test artscomponents/radar/TestIyActive_wfuns.arts
This test fails because of the change from iyActiveSingleScat2
to iyActiveSingleScat, and removal of the old code.  Chalmers
has taken it upon themselves to have a look and fix this test.

The differences are mostly in the last element of the derivatives.
All others are more or less the same but the last element seems to
be handled wrong in the new iyActiveSingleScat.  I am probably to
blame for this, having likely misunderstood some indexing.  If so,
maybe even the much smaller differences in the other elements that
manifest between these versions can be minimized.

WARNING:  This commit should be either reversed or purged before
we are happy with the implementation.