Basic debug error changes in ARTS
Introduces macros:

* ARTS_NOEXCEPT
* ARTS_CHECK(condition)
* ARTS_CHECK_WITH_MESSAGE(condition, ...)
* ARTS_ASSERT(condition)
* ARTS_ASSERT_WITH_MESSAGE(condition, ...)

The first is meant to be used instead of noexcept on functions
where one of the other four are being used to check the function
input.  The others throws a runtime error if the condition is bad.
Check-macros require a true statement to throw and assert-macros
requires a false statement to throw.  The variadic arguments
parses the input using (os << ... << args), to create an additional
message for the output of the runtime_error

What other changes do we want?  Do we want some way for the user to
turn off user errors?  That is so functions like zeeman_on_the_fly,
which has 19 different checks will fail at a lower level or just
produce absolute nonsense if any of the input is wrong?  These
Workspace methods that are mostly used inside Agendas of course
have all of these errors checked several times during a run, since
they can be called quite ofter.  I am sure some small amount of time
can be saved if they are made to never throw, so to speak, with some
simple -DNO_ARTS_USER_ERRORS=1 cmake flag.