Update to Eigen 3.3.9 (December 2020)
Also add curve_fit similar to python scipy.optimize.curve_fit
but with a little bit different interface

This addition required adding the unsupported/ modules of Eigen
if this is not desired, I have a port of the underlying MinPack
code that accepts a mostly similar interface.  I prefer using the
Eigen interface since it is possible to read and understand instead
of relying heavily on pointers inside lapack/blas functions

I cannot figure out how to use invlib for the task of minimizing these
things, and it seems to not be too important to do so since the resulting
output is actually not very easy to use (the Rosenkranz paramters don't
seem to actually work very well when coming from ECS).  Still, the
method to confirm that is tedious without this code, so the code must be
available somehow...