Updates m_abs.cc and some of its dependencies to use new error macro
This was done using [[maybe_unused]] on variables
that can be ignored in reckless mode.  Are we OK
with instead having build time warnings?  I find
it an additional neat way to mark that a variable
is just around as debug and doesn't do anything
'important'.

Note that sometimes this update was not straightforward:

In abs_xsec_per_speciesAddConts, there is an error
that lists everything in three lists by effectively
not using our list (Array<>, Vector) output operators.
This makes it impossible to use the ARTS_USER_ERROR_IF.
I choose to still change it to use ARTS_USER_ERROR_IF
because it is the only place where we ignore our own
output operators.  Have a look at this change and
revert if needed.

In propmat_clearskyAddFromAbsCoefPerSpecies, one error
checking is nested deep into some series of ifs inside
a for-loop.  Personally, I think this should be separated
out so the error checking does not happen inside a hot-
loop but I left the logic as it was.  This was probably
my addition a long time ago but I don't want to go in and
change working logic...

In propmat_clearskyAddFaraday, there is a lot of logic
involved with the error checking.  Often, the code reads
if () throw else do_something().  I choose to simply
remove the else statement and change code indention