Make Jacobian target store keys in usable manners
Also add a few more Options to constants.h for dealing with Jacobian.
Some of these had spaces in them before.  These are now removed since
enums cannot have spaces in their names.  Would it be better to allow
spaces again?  I think it's fine not to, and it would require a few
more steps to allow this.