Fix codesign issue on M1 Macs
On M1 Macs, the Python process gets killed by the OS when importing
pyarts after recompiling the libarts_api library.

macOS codesigns libraries and executables on first execution.  This
signature is verified on each further execution to ensure the binary has
not been tampered with.  The file is not allowed to be changed and
re-signed.  The inode of the file has to change to be able to generate a
new signature for the file.

shutil.copy, used in setup.py to overwrite the old library with the
recompiled version, replaces the contents of the old file without
changing its inode.  This leads to a mismatched signature on the next
import of the library into Python.

This problem is avoided by deleting the old library before copying the
new one into pyarts/workspace.