Remove In() and ilks from QuantumIdentifier
This unifies the matching of the lines.

I need next fix the existence of "do_vmr_jacobian".  Its
entire existence is wrong.  It just looks at the line but
the point of the LBL jacobian for VMR is to also allow for
broadeners and likewise to act upon the data.  This is
disallowed by how the current function does things.  I
need to remove it and handle VMR derivatives more
generally...