Add xml load/save support for ARTS C++ classes
Add compatibility to the xml load/save PyARTS functions for ARTS
classes that are only available through the C++ API.

load first tries to read the file with the Python interface.
If that fails, it uses the C++ classes interface.

save checks if the variable contains a savexml function which all the C++
ARTS classes do. If not, it falls back to the Python XML writer.