Add options to the Lbl sparse speedup
Three options exist:
None (Do everything)
LinearEven (Use a linear interpolation and 'cleverly' adjust on-the-fly)
QuadraticIndependent (Use a triplet of independent grids for each
                      sparse grid point)

These need more testing.  For 2.5 million frequencies they are about
the same speed.  For 25000, the Linear is twice as fast as the
quadratic, which is still twice as fast as the full calcs.

I think there might be some errors here somewhere but I need to take
some time to figure it out...