Make LinearEven option into LinearIndependent and follow the same scheme as QuadraticIndependent, i.e., that each range is independent
This doubles the number of calls to the core line shape but moves
all interpolation out of the core routine.  The QuadraticIndependent
method is still the preferred one, since 1/f^2 is approximately the
Lorentz lineshape, which all line shapes approximates far from their
line centers.

Anyways, the linear option for sparsification now works and it
can potentially be more than 50% faster than the quadratic
sparsification.

Note that boith methods suffers from recalculations of the inter-
section frequencies instead of just copying the values.  I don't
know how to solve this in a manageable way