Ignore single extra points in the sparse range
If you have only values at f2, the interpolation between f1 and f0
will be negative (the same is true if f0 and f2 are switched) in
quad.  Thus there must be 2 or 3 sparse grid points in the outer-
most ranges of the sparse grid to guarantee positive values in the
quad interpolation.  This means that there will be some values that
may be 0 between the true cut-off and the outer-most sparse-limit