Remove all of Linefunctions namespace.
This is fully replaceable by the LineShape namespace functionality,
with added benefits of simpler, faster, and accurate calculations.

This removes the ability to use abs_xsec_per_speciesAddLines for
jacobian calculations.

In fact, I don't want people to use abs_xsec_per_speciesAddLines anymore,
so I added depr.h to print a warning to std::cerr about its deprecation.
I want the function removed once the lookup-table adaptation is done,
but I don't know who is doing that or when it is ready...

This is a temporary solution for deprecating functions.  It would be
better with a more method-oriented (i.e., on parsing rather than execution)
warning system.  But depr.h macro can be used without much problem elsewhere
as well.