Small update to lineshapes so that the full thing compiles on new work computer
The update to dy.xml is necessary for unknown reasons, the FVC derivative has
changed.

Test GUI links quite OK on linux but it doesn't link on mac, so it is also updated.

The main thing here is that you can now get to the line shape calculations with
much less effort outside of lineshape.cc.  The variant-ism is now shifted inside
of a class that wraps the expected behavior