Remove the multiply-then-divide idiosyncrasy for Jacobian::Special::ArrayOfSpeciesTagVMR
This means that derivatives should now work for 0-VMR

The problem was that there used to be an adaptation in
adapt_stepwise_partial_derivatives so that these derivatives were rescaled by VMR.

This means that they had used to have to be output of the propagation matrix agenda
first multiplied by the VMR, so that the adaptation's divide by VMR would not
mess up the values.  This behavior was kept to keep the behavior to pre-analytical
Jacobian calculations identical to how they were back then.

Since we want to use the propagation matrix agenda to compute the lookup table,
it is now desired that the 0-VMR test case works.  This means that the old
behavior can no longer be kept.