Fix full line mixing for CO2 (seems to be working, but some questions remain)
The main algorithm is simplified and looks more similar to the O2 method now.

There are still small differences to the HITRAN port.  Mainly though in the
very far wing so absorption values are low.  The full algorithm seems to have
fewer negative absorption frequency ranges, though I can still spot a few in
my tests.  I am not sure how to deal with these...

Also fixed test_gui so that it works again (in this branch)