Add derivatives
This adds derivatives.  I had a huge problem here with the self model derivative
until I understood that everything was correct.  The self model depends on
VMR-squared, so of course the "for_species_tag=1" derivatives (e.g., old ARTS
style) will fail to give good results.  They are now manipulated to give the
same results as they would had the model been in the legacy code.

I hope no one is going to use or have used any of the models in legacy_continua.cc
to derive VMR from data because it is clearly a very bad idea...