Finished function ext_abs_pfun_from_tro.
The function is used by WSM HydrotableCalc and the DISORT WSMs. In
the case of DISORT, the new function can be selected by GIN only_tro.
If your scat_data just contain TRO, the new option gives a speed improvement
in the order of 25%, but depends on nstreams and num,ber of scattering
elements. Highest improvement for low nstreams and many scattering elements.