Make gen_auto_py build withoug needing artscore
This moves some stuff around to solve the problem of dependencies above.
It has one key drawback: The default value of workspace variables cannot
be printed anymore.  It can still be accessed on an empty variable but
it cannot be printed.  I am not sure if this can be fixed or if there is
a way to be self-referential in pyarts documentation somehow (e.g., access
the variable some way while building the string that becomes the
documentation)