Add several matrix+vector operations to Arts for matpack
You can now do most simple math directly on matpack types.  In fact,
if A, B and C are matrices of the correct sizes and x, y and z are vectors
of the right sizes, you can do (some of these operations might resize)
z = A * x + y;
row_vec(z).noalias() = A * x;  // This does not allocate any memory
B = A * A;
mat(B).noalias() = A * A;  // This does not allocate any memory
Matrix D(A * A + B * B - C);
Vector a(A * x + z - y);

An many more combinations.  At least these should all work

This also fixes a problem with assignment and constructor from
a matpack-like type in matpackI.h where the type cannot be an
actual matpack type in the template