Fix memory leak caused by cyclic reference
Workspace was not deleted if it contained an Agenda variable.
Since Agenda had a shared_ptr to the workspace it was living in,
it was creating a cyclic reference between the Agenda and the Workspace.
Thus, the Workspace destructor was never called.
This affected PyARTS when using multiple Workspaces.

Thanks to Paulina and Robert for reporting this bug.