Redo the CIA test
The "band" that the CIA is gridded on does not look like a band.
The new interpolation routine produces different value than the
old calculations.  However, these differences were only in what
is practically a Dirac absorption in the CIA "band".  I have
optimized the performance so it should be faster, and this changed
the central bin of the interpolation by at most 1 position so
that upwards movements don't move back when the match is less than
what the downwards motion wants to compensate for... i.e., a >
is now a >=, halving the set up time in some common cases.

This causes problems at a Dirac absorption here.

I update the results to just be in the center of the CIA and reran
with the old interpolation routine to generate a new absorption
spectra.  This is the new comparison XML file that is used in here