Add comments about input constants.
Also add "parsafe" option with a default "1".  This flag, if it
evaluates to true, makes the code safe for use in parallel code.
The cost is just a 3X slowdown.  On the one hand, it is sad it
is slower to be this way and I am considering fixing that with
a way to extract field data for wind and mag instead of component
data.  On the other hand, we have to have several times more calls
to the core computations in today's system with larger errors
in the actual field so I think this is still an improvement