Remove temperature field
Water equivalent agenda is not working properly after this.
There must have been a mistake in its design.  In all of the
OEM part of the code, there exist no fields so they must be
manually computed instead.  A better design would perhaps be
that it only takes a single temperature or (now) a single
atmospheric point.  Perhaps more so, should it just be part
of the atmospheric field?  Anyways, it has to be fixed because
it was not used properly before and this refactor shows it.