Fix TestIyActive_wfuns
Fix bug where dpnd_field_dx was not properly initialized.

As proposed by Patrick, the error message in iyRadarSingleScat about the
possible issues with the newly implemented jacobians has been removed,
assuming that the results are valid.

TestIyActive_wfuns compared the analytical against the perturbed
jacobians. With the current version this does not only result in large relative
differences at the top of the atmosphere, but also in a few other places in the
atmosphere. These other differences are always in places without a
signal, where the jacobian are very close to zero. Therefore, instead of
comparing analytical vs. perturbed jacobians, the analytical jacobian
results are now compared against newly added reference files.

Since the analytical jacobians are considered more accurate and the
recommended way to calculate jacobians, Richard and me decided to remove
the perturbed calculations from the test case.