Fix abs_lookupSetupWide internal vmr
SetupWide uses an assumed VMR for H2O to calculate the absorption
coefficients before converting them back to absorption cross sections.
Due to non-linearity, the assumed constant VMR for H2O can lead to
inaccurate crosssection values in some cases.
The constant VMR was replaced with a better approximation depending on
pressure.
The parameters for this approximation are also available as input
parameters of the WSM for further tweaking if needed.

Co-authored-by: Manfred Brath <manfred.brath@uni-hamburg.de>